backend/requirements.txt
asgiref==3.9.1
certifi==2025.8.3
chardet==3.0.4
charset-normalizer==3.4.3
Django==5.2.5
django-cors-headers==4.7.0
djangorestframework==3.16.1
djangorestframework_simplejwt==5.5.1
filelock==3.18.0
fsspec==2025.7.0
googletrans==4.0.0rc1
gunicorn==23.0.0
h11==0.9.0
h2==3.2.0
hf-xet==1.1.7
hpack==3.0.0
hstspreload==2025.1.1
httpcore==0.9.1
httpx==0.13.3
huggingface-hub==0.34.4
hyperframe==5.2.0
idna==2.10
Jinja2==3.1.6
joblib==1.5.1
MarkupSafe==3.0.2
mpmath==1.3.0
networkx==3.5
numpy==2.3.2
packaging==25.0
pillow==11.3.0
PyJWT==2.10.1
PyMuPDF==1.26.3
python-dotenv==1.1.1
python-magic==0.4.27
PyYAML==6.0.2
regex==2025.7.34
requests==2.32.4
rfc3986==1.5.0
safetensors==0.6.2
scikit-learn==1.7.1
scipy==1.16.1
sniffio==1.3.1
sqlparse==0.5.3
stripe==12.4.0
sympy==1.14.0
threadpoolctl==3.6.0
tokenizers==0.21.4
torch==2.8.0
tqdm==4.67.1
transformers==4.55.0
typing_extensions==4.14.1
urllib3==2.5.0
